"finished"** new controller. added post office greybox
The new controller is in a more or less complete state. It can be built upon and improved of course and I expect to make further changes later down the line.

The post office is a mockup of what a finished one could look like. I will use this scene to develop the level selection and dialogue systems until we get a finished version.